# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODC Pratique API",
        "description": "API documentation for ODC Pratique project",
        "contact": {
            "email": "votre@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Obtenir la liste des catégories",
                "operationId": "94e034e2bdb507c5e970d49c34fdc846",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS",
                                                "ERROR"
                                            ]
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Categorie"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "c7ed1b35eb40d77126f20ff47615f09a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS",
                                                "ERROR"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Categorie"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Obtenir les informations d'une catégorie",
                "operationId": "46c5c3954df07301e16ee08eaf957b86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS",
                                                "ERROR"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Categorie"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "7b13d3ed60e6e7337941a54b21084983",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS",
                                                "ERROR"
                                            ]
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie existante",
                "operationId": "d8a5823058e769ba7c26d6afd746f73e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "SUCCESS",
                                                "ERROR"
                                            ]
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Categorie"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categorie": {
                "required": [
                    "nom"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Électronique"
                    },
                    "description": {
                        "type": "string",
                        "example": "Produits électroniques et accessoires",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}